requirements.txt
astroid==2.2.5
certifi==2019.6.16
colorama==0.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.8
django-bootstrap-modal-forms==1.4.1
django-chartjs==1.5.0
django-widget-tweaks==1.4.2
djangorestframework==3.10.3
gunicorn==19.9.0
html5lib==1.0.1
isort==4.3.20
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pdfkit==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
PyPDF2==1.26.0
python-decouple==3.1
pytz==2018.5
reportlab==3.5.26
selenium==3.14.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.0
urllib3==1.25.3
webencodings==0.5.1
whitenoise==4.1.4
wincertstore==0.2
wkhtmltopdf==0.2
wrapt==1.11.2
xhtml2pdf==0.2.3
